+++ netty-loom-spring-boot-starter/src/test/java/io/github/azholdaspaev/nettyloomspring/autoconfigure/forward/app/ForwardTestApplication.java
@@ -1,0 +1,7 @@
+package io.github.azholdaspaev.nettyloomspring.autoconfigure.forward.app;
+
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class ForwardTestApplication {
+}